# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ferelix Server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/playback-info/{media_id}": {
      "post": {
        "tags": [
          "media"
        ],
        "summary": "Get Playback Info",
        "description": "Get playback information for a media file.\n\nAnalyzes the media file against the provided device profile to determine\nthe optimal playback method (DirectPlay, DirectStream, or Transcode).\n\nArgs:\n    media_id: ID of the media file\n    playback_request: Device profile and playback preferences\n    _user: Authenticated user (dependency)\n    session: Database session\n\nReturns:\n    Playback information with recommended streaming method\n\nRaises:\n    404: Media file not found",
        "operationId": "get_playback_info_api_v1_playback_info__media_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Media Id"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybackInfoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/libraries": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Libraries",
        "description": "Get enabled library paths (authenticated users).\n\nReturns only enabled libraries. Admin users should use /dashboard/libraries\nto see all libraries including disabled ones.\n\nArgs:\n    _user: Authenticated user (dependency)\n    session: Database session\n\nReturns:\n    List of enabled library paths",
        "operationId": "get_libraries_api_v1_libraries_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibrarySchema"
                  },
                  "title": "Response Get Libraries Api V1 Libraries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/libraries/{library_id}/items": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Library Items",
        "description": "Get items (media files) from a specific library.\n\nArgs:\n    library_id: Library path ID\n    user: Authenticated user\n    session: Database session\n    skip: Number of records to skip\n    limit: Maximum number of records to return\n\nReturns:\n    List of media files from the library\n\nRaises:\n    HTTPException: If library not found",
        "operationId": "get_library_items_api_v1_libraries__library_id__items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Library Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaFileSchema"
                  },
                  "title": "Response Get Library Items Api V1 Libraries  Library Id  Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media-files": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Media Files",
        "description": "Get all discovered media files.\n\nArgs:\n    user: Authenticated user\n    session: Database session\n    skip: Number of records to skip\n    limit: Maximum number of records to return\n\nReturns:\n    List of media files",
        "operationId": "get_media_files_api_v1_media_files_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaFileSchema"
                  },
                  "title": "Response Get Media Files Api V1 Media Files Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media-files/continue-watching": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Continue Watching Media Files",
        "description": "Get media files started by the current user but not completed.",
        "operationId": "get_continue_watching_media_files_api_v1_media_files_continue_watching_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 24,
              "title": "Limit"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaFileSchema"
                  },
                  "title": "Response Get Continue Watching Media Files Api V1 Media Files Continue Watching Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaybackInfoRequest": {
        "properties": {
          "DeviceProfile": {
            "$ref": "#/components/schemas/DeviceProfile"
          },
          "EnableDirectPlay": {
            "type": "boolean",
            "title": "Enabledirectplay",
            "default": true
          },
          "EnableDirectStream": {
            "type": "boolean",
            "title": "Enabledirectstream",
            "default": true
          },
          "EnableTranscoding": {
            "type": "boolean",
            "title": "Enabletranscoding",
            "default": true
          },
          "AllowVideoStreamCopy": {
            "type": "boolean",
            "title": "Allowvideostreamcopy",
            "default": true
          },
          "AllowAudioStreamCopy": {
            "type": "boolean",
            "title": "Allowaudiostreamcopy",
            "default": true
          },
          "IsPlayback": {
            "type": "boolean",
            "title": "Isplayback",
            "default": true
          },
          "RequestedResolution": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestedresolution"
          }
        },
        "type": "object",
        "required": [
          "DeviceProfile"
        ],
        "title": "PlaybackInfoRequest",
        "description": "Request for playback information."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}